% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlm.R
\name{AICR.Huber}
\alias{AICR.Huber}
\title{Robust Akaike information criterion for Huber M-estimation}
\usage{
AICR.Huber(X, y, beta, scale, k2 = 1.345)
}
\arguments{
\item{X}{numeric design matrix.}

\item{y}{numeric response vector.}

\item{beta}{numeric vector of regression coefficients from a Huber
M-estimation fit.}

\item{scale}{scale estimate from a Huber M-estimation fit.}

\item{k2}{tuning constant for the Huber \ifelse{latex}{\out{$\mathit{\psi}$}}{\ifelse{html}{\out{<i>&psi;</i>}}{*psi*}}-function. Default = \code{1.345}.}
}
\value{
The AICR value.
}
\description{
Computes Ronchetti's robust analogue¹ of the Akaike information criterion²
(AICR) for a linear model fitted by Huber M-estimation³. The effective
degrees of freedom term is evaluated using the the Moore–Penrose
pseudoinverse⁴ for numerical stability under rank deficiency.
}
\examples{
y <- c(36.3, 47.9, 47.2, 43.9, 47.6, 49.6, 53.2, 59.3, 63.2, 70.8, 75.9, 88.5,
       97.3, 103.6, 6.1, 120.2, 135.8, 139.4)
x <- as.matrix(1:length(y) - 1)
fit <- rlm.Huber(X = x, y = y)
print(AICR.Huber(x, y, fit$coefficients, fit$s))
}
\references{
\enumerate{
\item Ronchetti, E., 1985. Robust model selection in regression. \emph{Statistics &
Probability Letters}, 3(1), pp. 21–23.
\item Akaike, H., 1974. A new look at the statistical model identification.
\emph{IEEE Transactions on Automatic Control}, 19(6), pp. 716–723.
\item Huber, P.J., 1973. Robust regression: asymptotics, conjectures and Monte
Carlo. \emph{The Annals of Statistics}, pp. 799–821.
\item Penrose, R., 1955. A generalized inverse for matrices. In: \emph{Mathematical
Proceedings of the Cambridge Philosophical Society}, 51(3), pp. 406–413.
Cambridge: Cambridge University Press.
}
}
