% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.delim.R
\name{OR.delim.intersect}
\alias{OR.delim.intersect}
\title{Check if two delimited strings share any element (case-insensitive and
whitespace-trimmed)}
\usage{
OR.delim.intersect(a, b, delimiter = ",")
}
\arguments{
\item{a}{String vector, matrix, or data frame.}

\item{b}{String vector, matrix, or data frame of the same shape as \code{a}.}

\item{delimiter}{String delimiter. Default = \code{","}.}
}
\value{
A logical vector or matrix indicating whether \code{a} and \code{b} share any element.
}
\description{
For each pair of corresponding strings in \code{a} and \code{b}, applies \code{OR.delim.split()}
and checks whether any element is shared between \code{a} and \code{b}.
}
\seealso{
Other delim: 
\code{\link{OR.delim.contains}()},
\code{\link{OR.delim.merge}()},
\code{\link{OR.delim.replace}()},
\code{\link{OR.delim.split}()},
\code{\link{OR.delim.subset}()},
\code{\link{OR.delim.table}()}
}
\concept{delim}
