% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.delim.R
\name{OR.delim.merge}
\alias{OR.delim.merge}
\title{Merge delimited string elements pairwise (case-insensitive and
whitespace-trimmed)}
\usage{
OR.delim.merge(x1, x2, delimiter = ",")
}
\arguments{
\item{x1}{String vector, matrix, or data frame.}

\item{x2}{String vector, matrix, or data frame of the same shape as \code{x1}.}

\item{delimiter}{String delimiter. Default = \code{","}.}
}
\value{
A string vector or matrix with merged delimited elements.
}
\description{
For each pair of corresponding strings in \code{x1} and \code{x2}, splits the strings by
a specified delimiter, trims whitespace, converts to lowercase, takes the
union of the delimited elements, removes duplicate elements, sorts elements
alphabetically, and rejoins the elements using the same delimiter.
}
\seealso{
Other delim: 
\code{\link{OR.delim.contains}()},
\code{\link{OR.delim.replace}()},
\code{\link{OR.delim.table}()}
}
\concept{delim}
