% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.delim.R
\name{OR.delim.replace}
\alias{OR.delim.replace}
\title{Replace delimited string elements (case- and whitespace-insensitive)}
\usage{
OR.delim.replace(x, match, replacement, delimiter = ",")
}
\arguments{
\item{x}{String vector, matrix, or data frame containing delimited strings.}

\item{match}{String vector of elements to match (case- and
whitespace-insensitive).}

\item{replacement}{String delimited element replacement (case- and
whitespace-insensitive).}

\item{delimiter}{String delimiter. Default = \code{","}.}
}
\value{
A string vector or matrix with replaced delimited elements.
}
\description{
Splits each string in a vector, matrix, or data frame by a specified
delimiter, trims whitespace, converts to lowercase, and checks whether any
delimited element matches one or more specified strings. Replaces any
matching elements with a new string, removes duplicate elements, sorts
elements alphabetically, and rejoins the elements using the same delimiter.
}
\examples{
treatment <- c("capecitabine",
               "LETROZOLE",
               "letrozole, palbociclib",
               "Letrozole,Ribociclib",
               "anastrozole, ribociclib")
treatment <- OR.delim.replace(treatment, "anastrozole", "ai")
treatment <- OR.delim.replace(treatment, "letrozole", "ai")
treatment <- OR.delim.replace(treatment, "palbociclib", "cdk46i")
treatment <- OR.delim.replace(treatment, "ribociclib", "cdk46i")
print(treatment)
}
