% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.delim.R
\name{OR.delim.split}
\alias{OR.delim.split}
\title{Split delimited string elements (case-insensitive and whitespace-trimmed)}
\usage{
OR.delim.split(x, delimiter = ",")
}
\arguments{
\item{x}{String vector.}

\item{delimiter}{String delimiter. Default = \code{","}.}
}
\value{
A character vector of delimited elements.
}
\description{
Splits a string by a specified delimiter, replaces \code{NA} with \code{""}, trims
whitespace, converts to lowercase, and returns delimited elements.
}
\seealso{
Other delim: 
\code{\link{OR.delim.contains}()},
\code{\link{OR.delim.intersect}()},
\code{\link{OR.delim.merge}()},
\code{\link{OR.delim.replace}()},
\code{\link{OR.delim.subset}()},
\code{\link{OR.delim.table}()},
\code{\link{OR.delim.txclass}()},
\code{\link{OR.delim.txname}()}
}
\concept{delim}
