% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.delim.R
\name{OR.delim.subset}
\alias{OR.delim.subset}
\title{Check if one delimited string is a subset of another (case-insensitive and
whitespace-trimmed)}
\usage{
OR.delim.subset(a, b, delimiter = ",")
}
\arguments{
\item{a}{String vector, matrix, or data frame.}

\item{b}{String vector, matrix, or data frame of the same shape as \code{a}.}

\item{delimiter}{String delimiter. Default = \code{","}.}
}
\value{
A logical vector or matrix indicating whether \code{b} is a subset of \code{a}.
}
\description{
For each pair of corresponding strings in \code{a} and \code{b}, applies \code{OR.delim.split()}
and checks whether all elements in \code{b} are
contained in \code{a}.
}
\seealso{
Other delim: 
\code{\link{OR.delim.contains}()},
\code{\link{OR.delim.intersect}()},
\code{\link{OR.delim.merge}()},
\code{\link{OR.delim.replace}()},
\code{\link{OR.delim.split}()},
\code{\link{OR.delim.table}()},
\code{\link{OR.delim.txclass}()},
\code{\link{OR.delim.txname}()}
}
\concept{delim}
