% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.delim.R
\name{OR.delim.txclass}
\alias{OR.delim.txclass}
\title{Classify common medical oncology treatments (work in progress)}
\usage{
OR.delim.txclass(x, delimiter = ",")
}
\arguments{
\item{x}{String vector of delimited treatments.}

\item{delimiter}{String delimiter. Default = \code{","}.}
}
\value{
A string vector of delimited treatment class labels.
}
\description{
For each string in \code{x}, applies \code{OR.delim.split()} and replaces common cancer
treatments with predefined treatment class ("endocrine", "chemotherapy",
"immunotherapy", or "targeted") and selected subclass (such as "her2 and
"pd1") labels. Multiple labels, together with any unclassified elements, are
returned as a delimited string.
}
\seealso{
Other delim: 
\code{\link{OR.delim.contains}()},
\code{\link{OR.delim.intersect}()},
\code{\link{OR.delim.merge}()},
\code{\link{OR.delim.replace}()},
\code{\link{OR.delim.split}()},
\code{\link{OR.delim.subset}()},
\code{\link{OR.delim.table}()},
\code{\link{OR.delim.txname}()}
}
\concept{delim}
