% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.delim.R
\name{OR.delim.txname}
\alias{OR.delim.txname}
\title{Standardize common medical oncology treatment names (work in progress)}
\usage{
OR.delim.txname(x, delimiter = ",")
}
\arguments{
\item{x}{String vector of delimited treatments.}

\item{delimiter}{String delimiter. Default = \code{","}.}
}
\value{
A string vector of delimited standarized treatment names.
}
\description{
For each string in \code{x}, standardizes common cancer treatment names:
\enumerate{
\item Removes administration routes (such as "iv" and "oral"), scheduling terms
(such as "q2w" and "weekly"), and other modifiers (such as "modified" and
"single agent").
\item Converts common separators ("&", "+", ",", "/", " and ", " with ") to a
consistent delimiter.
\item Removes supportive agents (such as "folinic acid").
\item Replaces common trade names,  abbreviations (such as "FOLFOX"), and
synonyms (such as "T-DM1") with standardized names.
}
}
\seealso{
Other delim: 
\code{\link{OR.delim.contains}()},
\code{\link{OR.delim.intersect}()},
\code{\link{OR.delim.merge}()},
\code{\link{OR.delim.replace}()},
\code{\link{OR.delim.split}()},
\code{\link{OR.delim.subset}()},
\code{\link{OR.delim.table}()},
\code{\link{OR.delim.txclass}()}
}
\concept{delim}
