% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.Google.R
\name{OR.Google.address}
\alias{OR.Google.address}
\title{Resolve address using Google Geocoding API}
\usage{
OR.Google.address(address, key, cache)
}
\arguments{
\item{address}{String input address.}

\item{key}{String Google API key.}

\item{cache}{String path to a \code{cachem} disk cache directory.}
}
\value{
String formatted address or \code{NA} if no match is found.
}
\description{
Returns the formatted address for a given input string. To reduce API calls,
addresses are case-insensitive and whitespace-trimmed, empty addresses after
processing return \code{NA} early, and results are memoized to disk.
}
\seealso{
Other GoogleMapsPlatform: 
\code{\link{OR.Google.TSP}()},
\code{\link{OR.Google.distance}()}
}
\concept{GoogleMapsPlatform}
