% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.Google.R
\name{OR.Google.distance}
\alias{OR.Google.distance}
\title{Road distance between two addresses using Google Routes API}
\usage{
OR.Google.distance(a, b, key, cache)
}
\arguments{
\item{a}{String address A.}

\item{b}{String address B.}

\item{key}{String Google API key.}

\item{cache}{String path to a \code{cachem} disk cache directory.}
}
\value{
Numeric distance in kilometres or \code{NA} if no route is found.
}
\description{
Returns the road distance in km between two addresses. To reduce API calls,
addresses are case-insensitive and whitespace-trimmed, empty addresses after
processing return \code{NA} early, processed addresses are internally sorted, and
results are memoized to disk.
}
\details{
If no route is found, input addresses are compared via Google Geocoding API;
identical resolved addresses return \code{0}, otherwise return \code{NA}.
}
\seealso{
Other GoogleMapsPlatform: 
\code{\link{OR.Google.TSP}()},
\code{\link{OR.Google.address}()}
}
\concept{GoogleMapsPlatform}
