% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.line.R
\name{OR.line.truncate}
\alias{OR.line.truncate}
\title{Truncate stop dates for lines of treatment}
\usage{
OR.line.truncate(start.dates, stop.dates, max.days = 168)
}
\arguments{
\item{start.dates}{Numeric matrix of treatment start dates, with one row per
patient and one column per line of treatment.}

\item{stop.dates}{Numeric matrix of treatment stop dates, with one row per
patient and one column per line of treatment.}

\item{max.days}{Maximum number of days allowed between each start and stop
date. Default = \code{168}.}
}
\value{
A matrix of truncated stop dates.
}
\description{
For each line of treatment, truncates the recorded stop date so that it does
not exceed a specified number of days after its start date.
}
\details{
An example use case is with oxaliplatin-containing chemotherapy regimens such
as FOLFOX where the oxaliplatin component is usually limited to no more than
24 weeks and time beyond this represents treatment with 5-fluorouracil only.
}
