% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.imputation.R
\name{OR.LOCF.cluster}
\alias{OR.LOCF.cluster}
\title{Last observation carried forward within clusters}
\usage{
OR.LOCF.cluster(x, id = NULL)
}
\arguments{
\item{x}{Vector of values for LOCF imputation.}

\item{id}{Optional vector of cluster identifiers. If \code{NULL}, all observations
are treated as a single cluster.}
}
\value{
Vector of same length as \code{x} with LOCF applied within each \code{id}.
}
\description{
Applies LOCF imputation to \code{x} within each \code{id}, assuming observations are
already ordered within each \code{id}.
}
\concept{imputation}
