% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.R
\name{OR.mean}
\alias{OR.mean}
\title{Arithmetic mean with NA removal or NA if no valid entries}
\usage{
OR.mean(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Maximum of non-\code{NA} elements or \code{NA} if none exist.
}
\description{
Equivalent to \code{mean(x, na.rm = TRUE)} except returns \code{NA} (rather than \code{NaN})
if all elements are \code{NA}.
}
