% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.R
\name{OR.min}
\alias{OR.min}
\title{Minimum with NA removal or NA if no valid entries}
\usage{
OR.min(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Minimum of non-\code{NA} elements or \code{NA} if none exist.
}
\description{
Equivalent to \code{min(x, na.rm = TRUE)} except returns \code{NA} if all elements are
\code{NA}.
}
