% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.summary.R
\name{OR.mode}
\alias{OR.mode}
\title{Statistical mode}
\usage{
OR.mode(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
A scalar containing the mode of \code{x}.
}
\description{
Returns the most common value in \code{x}.
}
\seealso{
Other summary: 
\code{\link{OR.max}()},
\code{\link{OR.mean}()},
\code{\link{OR.min}()},
\code{\link{OR.modefreq}()},
\code{\link{OR.rowleft}()},
\code{\link{OR.rowmax}()},
\code{\link{OR.rowmax.threshold}()},
\code{\link{OR.rowmin}()},
\code{\link{OR.rowmin.threshold}()},
\code{\link{OR.sum}()}
}
\concept{summary}
