% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.summary.R
\name{OR.modefreq}
\alias{OR.modefreq}
\title{Frequency of the mode}
\usage{
OR.modefreq(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
Integer scalar giving the frequency of the mode.
}
\description{
Returns the number of occurrences of the most common value in \code{x}.
}
\seealso{
Other summary: 
\code{\link{OR.max}()},
\code{\link{OR.mean}()},
\code{\link{OR.min}()},
\code{\link{OR.mode}()},
\code{\link{OR.rowleft}()},
\code{\link{OR.rowmax}()},
\code{\link{OR.rowmax.threshold}()},
\code{\link{OR.rowmin}()},
\code{\link{OR.rowmin.threshold}()},
\code{\link{OR.sum}()}
}
\concept{summary}
