% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.R
\name{OR.NA.to.empty}
\alias{OR.NA.to.empty}
\title{Convert missing values to empty strings}
\usage{
OR.NA.to.empty(x)
}
\arguments{
\item{x}{String vector or matrix.}
}
\value{
The input object with all \code{NA} values replaced by \code{""}.
}
\description{
Converts all \code{NA} elements in a vector or matrix to \code{""}.
}
\seealso{
Other basic: 
\code{\link{OR.0.rm}()},
\code{\link{OR.F.to.NA}()},
\code{\link{OR.NA.rm}()},
\code{\link{OR.NA.to.0}()},
\code{\link{OR.NA.to.F}()},
\code{\link{OR.as.numeric}()},
\code{\link{OR.check.equal}()},
\code{\link{OR.string.to.binary}()}
}
\concept{basic}
