% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.outliers.R
\name{OR.outliers.rlm.plot}
\alias{OR.outliers.rlm.plot}
\title{Polynomial linear non-parametric outlier plot}
\usage{
OR.outliers.rlm.plot(
  x,
  y,
  max.degree = 3,
  p = 0.05,
  echo = FALSE,
  x.breaks = NA,
  x.labels = NA,
  x.title = "",
  y.min = NA,
  y.max = NA,
  y.title = ""
)
}
\arguments{
\item{x}{Numeric predictor vector.}

\item{y}{Numeric response vector.}

\item{max.degree}{Maximum polynomial degree considered. Default = \code{3}.}

\item{p}{Target two-sided exclusion proportion under normality for the
residual-based modified Z-score rule. Default = \code{0.05}.}
}
\description{
Fits a univariable polynomial regression using Huber M-estimation with
degree selected by robust AIC (AICR), then flags observations whose
residuals exceed a small-sample–adjusted multiple of the median absolute
deviation.
}
\examples{
y <- c(36.3, 47.9, 47.2, 43.9, 47.6, 49.6, 53.2, 59.3, 63.2, 70.8, 75.9, 88.5,
       97.3, 103.6, 6.1, 120.2, 135.8, 139.4)
x <- 1:length(y) - 1
OR.outliers.rlm.plot(x, y, max.degree = 3, p = 0.01, y.min = 0, y.max = 150)
}
