% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.other.R
\name{OR.read.csv.stub}
\alias{OR.read.csv.stub}
\title{Read CSV files matching a path stub and combine by rows}
\usage{
OR.read.csv.stub(path.stub, ...)
}
\arguments{
\item{path.stub}{String path prefix used to match CSV files.}

\item{...}{Additional arguments passed to \code{read.csv()}.}
}
\value{
A \code{data.frame} containing contents of all matched CSV files combined
by rows.
}
\description{
Reads all comma-separated values (CSV) files whose paths begin with \code{path.stub}
and combines them by rows.
}
\seealso{
Other other: 
\code{\link{OR.collapse}()},
\code{\link{OR.date.Excel}()},
\code{\link{OR.holdout}()},
\code{\link{OR.permutations}()},
\code{\link{OR.time.Excel}()}
}
\concept{other}
