% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.regimen.R
\name{OR.regimen.truncate}
\alias{OR.regimen.truncate}
\title{Truncate stop dates by maximum duration}
\usage{
OR.regimen.truncate(date_start, date_stop, t_max)
}
\arguments{
\item{date_start}{Numeric matrix of start dates.}

\item{date_stop}{Numeric matrix of stop dates.}

\item{t_max}{Maximum allowed duration.}
}
\value{
Numeric matrix of truncated stop dates where
\code{date_stop - date_start} does not exceed \code{t_max}.
}
\description{
Limits the duration between \code{date_start} and \code{date_stop} to \code{t_max} by
truncating \code{date_stop} where necessary.
}
