% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.R
\name{OR.rowleft}
\alias{OR.rowleft}
\title{Row-wise leftmost non-missing value}
\usage{
OR.rowleft(input)
}
\arguments{
\item{input}{A matrix or data frame.}
}
\value{
A vector of the leftmost non-missing values by row.
}
\description{
Returns the leftmost non-missing value in each row of a matrix or data frame.
Returns \code{NA} if an entire row is missing.
}
\details{
An example use case is single imputation where column order defines the
hierarchy of preference, such as prioritizing pathological stage over
clinical stage.
}
