% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.summary.R
\name{OR.rowmax}
\alias{OR.rowmax}
\title{Row-wise maxima}
\usage{
OR.rowmax(input)
}
\arguments{
\item{input}{A numeric matrix or data frame.}
}
\value{
A numeric vector of row-wise maxima.
}
\description{
Computes the maximum value of each row in a matrix or data frame, ignoring
\code{NA}s. Returns \code{NA} if an entire row is missing.
}
\details{
An example use case in survival analysis is determining the date of last
follow-up from several dates when the patient was observed.
}
\seealso{
Other summary: 
\code{\link{OR.max}()},
\code{\link{OR.mean}()},
\code{\link{OR.min}()},
\code{\link{OR.mode}()},
\code{\link{OR.modefreq}()},
\code{\link{OR.rowleft}()},
\code{\link{OR.rowmax.threshold}()},
\code{\link{OR.rowmin}()},
\code{\link{OR.rowmin.threshold}()},
\code{\link{OR.sum}()}
}
\concept{summary}
