% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.summary.R
\name{OR.rowmax.threshold}
\alias{OR.rowmax.threshold}
\title{Row-wise maxima below a threshold}
\usage{
OR.rowmax.threshold(input, threshold)
}
\arguments{
\item{input}{Numeric matrix or data frame.}

\item{threshold}{Numeric vector of thresholds of length equal to the number
of \code{input} rows.}
}
\value{
Numeric vector of row-wise maxima below a threshold.
}
\description{
For each row of a matrix or data frame, returns the maximum value strictly
less than a corresponding threshold, ignoring \code{NA}s. Returns \code{NA} if no
values satisfy the condition.
}
\seealso{
Other summary: 
\code{\link{OR.max}()},
\code{\link{OR.mean}()},
\code{\link{OR.min}()},
\code{\link{OR.mode}()},
\code{\link{OR.modefreq}()},
\code{\link{OR.rowleft}()},
\code{\link{OR.rowmax}()},
\code{\link{OR.rowmin}()},
\code{\link{OR.rowmin.threshold}()},
\code{\link{OR.sum}()}
}
\concept{summary}
