% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.summary.R
\name{OR.rowmin}
\alias{OR.rowmin}
\title{Row-wise minima}
\usage{
OR.rowmin(input)
}
\arguments{
\item{input}{A numeric matrix or data frame.}
}
\value{
A numeric vector of row-wise minima.
}
\description{
Computes the minimum value of each row in a matrix or data frame, ignoring
\code{NA}s. Returns \code{NA} if an entire row is missing.
}
\details{
An example use case in survival analysis is determining the event date for
progression-free survival based on the dates of progression and death.
}
\seealso{
Other summary: 
\code{\link{OR.max}()},
\code{\link{OR.mean}()},
\code{\link{OR.min}()},
\code{\link{OR.mode}()},
\code{\link{OR.modefreq}()},
\code{\link{OR.rowleft}()},
\code{\link{OR.rowmax}()},
\code{\link{OR.rowmax.threshold}()},
\code{\link{OR.rowmin.threshold}()},
\code{\link{OR.sum}()}
}
\concept{summary}
