% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.Excel.R
\name{OR.time.Excel}
\alias{OR.time.Excel}
\title{Convert mixed format times of day to Microsoft Excel serial times}
\usage{
OR.time.Excel(x)
}
\arguments{
\item{x}{String vector of times of day in serial or 24-hour clock format.}
}
\value{
Numeric vector of serial times using the Microsoft Excel convention.
}
\description{
Converts string times of day that may be in serial (using the Microsoft
Excel convention of a fraction of a 24-hour day) or 24-hour clock format
with (e.g., "12:00:00") or without (e.g., "12:00") seconds into numeric
serial times.
}
\examples{
print(OR.time.Excel(c("0.5", "12:00", "12:00:00")))
}
\seealso{
Other other: 
\code{\link{OR.collapse}()},
\code{\link{OR.date.Excel}()},
\code{\link{OR.holdout}()},
\code{\link{OR.permutations}()},
\code{\link{OR.read.csv.stub}()}
}
\concept{other}
