% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.txlines.timeline.R
\name{OR.txlines.timeline}
\alias{OR.txlines.timeline}
\title{Construct treatment timelines (long format)}
\usage{
OR.txlines.timeline(
  id,
  regimens,
  startdates,
  stopdates,
  stopreasons = NULL,
  eventdates = NULL,
  echo = FALSE
)
}
\arguments{
\item{id}{Vector of identifiers.}

\item{regimens}{Comma-delimited string matrix or data frame of treatment
regimens (one row per \code{id}, one column per line of treatment).}

\item{startdates}{Numeric matrix or data frame of start dates (one row per
\code{id}, one column per line of treatment).}

\item{stopdates}{Numeric matrix or data frame of stop dates (one row per \code{id},
one column per line of treatment).}

\item{stopreasons}{Optional comma-delimited string matrix or data frame of
stop reasons (one row per \code{id}, one column per line of treatment).}

\item{eventdates}{Optional numeric matrix or data frame of event dates (one
row per \code{id}, one column per line of treatment).}

\item{echo}{Logical. If \code{TRUE}, prints additional information to console.
Default = \code{FALSE}.}
}
\value{
A data frame in long format with columns:
\describe{
\item{id}{Identifier.}
\item{date}{Date when a treatment starts or stops, or an event occurs.}
\item{regimen}{Active regimens at the date.}
\item{stopreasons}{Stop reasons occurring at the date.}
\item{event}{Logical indicating whether an event occurs at the date.}
}
}
\description{
For each \code{id}, takes treatment lines stored in wide format (one column per
line of treatment for each of treatment regimens, start dates, stop dates,
stop reasons, and event dates) and constructs a time-ordered timeline in long
format. Each output row corresponds to a unique date at which any treatment
starts or stops, or an event (for example, disease progression) occurs. At
each unique date, active treatment regimens and stop reasons are merged.
}
\seealso{
Other txlines: 
\code{\link{OR.txlines.restart}()},
\code{\link{OR.txlines.truncate}()},
\code{\link{OR.txlines.wide}()}
}
\concept{txlines}
