% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.txlines.timeline.R
\name{OR.txlines.wide}
\alias{OR.txlines.wide}
\title{Convert treatment timelines to wide format}
\usage{
OR.txlines.wide(id, date, regimen, event, line)
}
\arguments{
\item{id}{Vector of identifiers.}

\item{date}{Numeric vector of dates (sorted within each \code{id}).}

\item{regimen}{Comma-delimited string vector of active regimens at each date.}

\item{event}{Logical vector indicating whether an event occurs at each date.}

\item{line}{Numeric vector indicating treatment line at each date.}
}
\value{
A data frame in wide format with columns, where \code{k} indexes treatment
lines:
\describe{
\item{id}{Identifier.}
\item{date.startL\if{html}{\out{<k>}}}{Start date of line \code{k}.}
\item{regimenL\if{html}{\out{<k>}}}{Regimens in line \code{k}.}
\item{date.eventL\if{html}{\out{<k>}}}{Earliest event date for line \code{k}.}
}
}
\description{
Takes treatment timelines in long format (as produced by
\code{OR.txlines.timeline()}) and converts them to wide format with one row per
\code{id} and one set of start date, regimen, and event date columns per treatment
line.
}
\seealso{
Other txlines: 
\code{\link{OR.txlines.restart}()},
\code{\link{OR.txlines.timeline}()},
\code{\link{OR.txlines.truncate}()}
}
\concept{txlines}
