% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.distributions.R
\name{rbeta.mm}
\alias{rbeta.mm}
\title{Generate beta random values from a mean and standard deviation}
\usage{
rbeta.mm(n, mu, sigma)
}
\arguments{
\item{n}{Number of random values to generate.}

\item{mu}{Mean \ifelse{latex}{\out{$\mathit{\mu}$}}{\ifelse{html}{\out{<i>&mu;</i>}}{*mu*}}.}

\item{sigma}{Standard deviation \ifelse{latex}{\out{$\mathit{\sigma}$}}{\ifelse{html}{\out{<i>&sigma;</i>}}{*sigma*}}.}
}
\value{
Numeric vector of length \code{n}.
}
\description{
Draws \code{n} random values from a beta distribution specified by its mean \ifelse{latex}{\out{$\mathit{\mu}$}}{\ifelse{html}{\out{<i>&mu;</i>}}{*mu*}}
and standard deviation \ifelse{latex}{\out{$\mathit{\sigma}$}}{\ifelse{html}{\out{<i>&sigma;</i>}}{*sigma*}},
rather than its shape parameters, via the method of moments.
}
\concept{distributions}
