% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlm.R
\name{rlm.Huber}
\alias{rlm.Huber}
\title{Robust linear regression using Huber M-estimation}
\usage{
rlm.Huber(X, y, k2 = 1.345, maxit = 100, tol = 1e-04)
}
\arguments{
\item{X}{numeric design matrix.}

\item{y}{numeric response vector.}

\item{k2}{tuning constant for Huber \ifelse{latex}{\out{$\mathit{\psi}$}}{\ifelse{html}{\out{<i>&psi;</i>}}{*psi*}}-function. Default = \code{1.345}.}

\item{maxit}{maximum number of iterations. Default = \code{100}.}

\item{tol}{convergence tolerance. Default = \code{0.0001}.}
}
\value{
An object of classes \code{"lm"} and \code{"rlm"}, broadly compatible with objects
returned by \code{MASS::rlm()} with minor differences due to rounding.
}
\description{
Fits a linear regression model using Huber M-estimation¹ via iteratively
reweighted least squares.
}
\references{
\enumerate{
\item Huber, P.J., 1973. Robust regression: asymptotics, conjectures and Monte
Carlo. \emph{The Annals of Statistics}, pp. 799–821.
}
}
