% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlm.R
\name{AICR.Huber}
\alias{AICR.Huber}
\title{Robust Akaike information criterion for Huber M-estimation}
\usage{
AICR.Huber(y, X, beta, scale, k2 = 1.345)
}
\arguments{
\item{y}{numeric response vector.}

\item{X}{numeric design matrix.}

\item{beta}{numeric vector of regression coefficients from a Huber
M-estimation fit.}

\item{scale}{scale estimate from the fit from a Huber M-estimation fit.}

\item{k2}{tuning constant for the Huber \ifelse{latex}{\out{$\mathit{\psi}$}}{\ifelse{html}{\out{<i>&psi;</i>}}{*psi*}}-function. Default = \code{1.345}.}
}
\value{
The robust AIC value.
}
\description{
Computes Ronchetti's robust analogue¹ of the Akaike information criterion²
(AICR) for a linear model fitted by Huber M-estimation³.
}
\references{
\enumerate{
\item Ronchetti, E., 1985. Robust model selection in regression. \emph{Statistics &
Probability Letters}, 3(1), pp. 21–23.
\item Akaike, H., 1974. A new look at the statistical model identification.
\emph{IEEE Transactions on Automatic Control}, 19(6), pp. 716–723.
\item Huber, P.J., 1973. Robust regression: asymptotics, conjectures and Monte
Carlo. \emph{The Annals of Statistics}, pp. 799–821.
}
}
