% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.R
\name{OR.0.rm}
\alias{OR.0.rm}
\title{Remove zero elements}
\usage{
OR.0.rm(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
\code{x} with all elements equal to \code{0} removed.
}
\description{
Removes elements equal to \code{0} from a numeric vector.
}
\seealso{
Other basic: 
\code{\link{OR.F.to.NA}()},
\code{\link{OR.NA.rm}()},
\code{\link{OR.NA.to.0}()},
\code{\link{OR.NA.to.F}()},
\code{\link{OR.NA.to.empty}()},
\code{\link{OR.as.numeric}()},
\code{\link{OR.string.to.binary}()}
}
\concept{basic}
