% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.R
\name{OR.0.rm}
\alias{OR.0.rm}
\title{Remove zero elements}
\usage{
OR.0.rm(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
\code{x} with all elements equal to \code{0} removed.
}
\description{
Removes elements equal to \code{0} from a numeric vector.
}
