% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.delim.R
\name{OR.delim.contains}
\alias{OR.delim.contains}
\title{Detect delimited string elements (case- and whitespace-insensitive)}
\usage{
OR.delim.contains(x, match, delimiter = ",")
}
\arguments{
\item{x}{String vector, matrix, or data frame containing delimited strings.}

\item{match}{String vector of elements to match (case- and
whitespace-insensitive).}

\item{delimiter}{String delimiter. Default = \code{","}.}
}
\value{
A logical vector or matrix indicating whether each string contains a
delimited element matching any of the specified strings.
}
\description{
delimiter, trims whitespace, converts to lowercase, and checks whether any
delimited element matches one or more specified strings.
}
