% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.delim.R
\name{OR.delim.merge}
\alias{OR.delim.merge}
\title{Merge delimited string elements pairwise (case-insensitive and
whitespace-trimmed)}
\usage{
OR.delim.merge(a, b, delimiter = ",")
}
\arguments{
\item{a}{String vector, matrix, or data frame.}

\item{b}{String vector, matrix, or data frame of the same shape as \code{a}.}

\item{delimiter}{String delimiter. Default = \code{","}.}
}
\value{
A string vector or matrix with merged delimited elements.
}
\description{
For each pair of corresponding strings in \code{a} and \code{b}, applies \code{OR.delim.split()},
takes the union of the delimited elements, removes duplicate elements, sorts
elements alphabetically, and rejoins the elements using the same delimiter.
}
\seealso{
Other delim: 
\code{\link{OR.delim.contains}()},
\code{\link{OR.delim.intersect}()},
\code{\link{OR.delim.replace}()},
\code{\link{OR.delim.split}()},
\code{\link{OR.delim.subset}()},
\code{\link{OR.delim.table}()},
\code{\link{OR.delim.txclass}()},
\code{\link{OR.delim.txname}()}
}
\concept{delim}
