% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.delim.R
\name{OR.delim.table}
\alias{OR.delim.table}
\title{Frequency table of delimited string elements (case- and
whitespace-insensitive)}
\usage{
OR.delim.table(x, delimiter = ",")
}
\arguments{
\item{x}{String vector, matrix, or data frame containing delimited strings.}

\item{delimiter}{String delimiter. Default = \code{","}.}
}
\value{
A table of delimited element frequencies (case- and
whitespace-insensitive).
}
\description{
Splits each string in a string vector, matrix, or data frame by a specified
delimiter, trims whitespace, converts to lowercase, and tabulates the
frequency of all unique delimited elements.
}
\examples{
treatment <- c("capecitabine",
               "LETROZOLE",
               "letrozole, palbociclib",
               "Letrozole,Ribociclib",
               "anastrozole, ribociclib")
print(OR.delim.table(treatment))
}
