% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.R
\name{OR.F.to.NA}
\alias{OR.F.to.NA}
\title{Convert FALSE values to NA}
\usage{
OR.F.to.NA(x)
}
\arguments{
\item{x}{Logical vector or matrix.}
}
\value{
The input object with all \code{FALSE} values replaced by \code{NA}.
}
\description{
Converts all \code{FALSE} elements in a vector or matrix to \code{NA}.
}
