% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.R
\name{OR.F.to.NA}
\alias{OR.F.to.NA}
\title{Convert FALSE values to NA}
\usage{
OR.F.to.NA(x)
}
\arguments{
\item{x}{Logical vector or matrix.}
}
\value{
The input object with all \code{FALSE} values replaced by \code{NA}.
}
\description{
Converts all \code{FALSE} elements in a vector or matrix to \code{NA}.
}
\seealso{
Other basic: 
\code{\link{OR.0.rm}()},
\code{\link{OR.NA.rm}()},
\code{\link{OR.NA.to.0}()},
\code{\link{OR.NA.to.F}()},
\code{\link{OR.NA.to.empty}()},
\code{\link{OR.as.numeric}()},
\code{\link{OR.string.to.binary}()}
}
\concept{basic}
