% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.Google.R
\name{OR.Google.TSP}
\alias{OR.Google.TSP}
\title{Travelling salesman distance via Google Routes API}
\usage{
OR.Google.TSP(origin, destinations, key, echo = FALSE)
}
\arguments{
\item{origin}{String starting and ending address.}

\item{destinations}{String vector of addresses to visit.}

\item{key}{String Google API key.}

\item{echo}{Boolean indicating whether to print progress and summary to
console. Default = \code{FALSE}.}
}
\value{
Numeric scalar shortest round-trip distance in km.
}
\description{
Computes the shortest round-trip distance in km starting and ending at the
origin, visiting each destination exactly once, via brute-force enumeration
of all permutations. Distances are computed via \code{OR.Google.distance()}.
}
\seealso{
Other GoogleMapsPlatform: 
\code{\link{OR.Google.address}()},
\code{\link{OR.Google.distance}()}
}
\concept{GoogleMapsPlatform}
