% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.R
\name{OR.max}
\alias{OR.max}
\title{Maximum with NA removal or NA if no valid entries}
\usage{
OR.max(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Maximum of non-\code{NA} elements or \code{NA} if none exist.
}
\description{
Equivalent to \code{max(x, na.rm = TRUE)} except returns \code{NA} (rather than \code{-Inf}) if
all elements are \code{NA}.
}
