% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.R
\name{OR.NA.rm}
\alias{OR.NA.rm}
\title{Remove missing values}
\usage{
OR.NA.rm(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
\code{x} with all \code{NA} entries removed.
}
\description{
Removes \code{NA} elements in a vector.
}
\seealso{
Other basic: 
\code{\link{OR.0.rm}()},
\code{\link{OR.F.to.NA}()},
\code{\link{OR.NA.to.0}()},
\code{\link{OR.NA.to.F}()},
\code{\link{OR.NA.to.empty}()},
\code{\link{OR.as.numeric}()},
\code{\link{OR.string.to.binary}()}
}
\concept{basic}
