% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.R
\name{OR.NA.rm}
\alias{OR.NA.rm}
\title{Remove missing values}
\usage{
OR.NA.rm(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
\code{x} with all \code{NA} entries removed.
}
\description{
Removes \code{NA} elements in a vector.
}
