% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.R
\name{OR.NA.to.0}
\alias{OR.NA.to.0}
\title{Convert missing values to zero}
\usage{
OR.NA.to.0(x)
}
\arguments{
\item{x}{Numeric vector or matrix.}
}
\value{
The input object with all \code{NA} values replaced by \code{0}.
}
\description{
Converts all \code{NA} elements in a vector or matrix to \code{0}.
}
