% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.R
\name{OR.NA.to.empty}
\alias{OR.NA.to.empty}
\title{Convert missing values to empty strings}
\usage{
OR.NA.to.empty(x)
}
\arguments{
\item{x}{String vector or matrix.}
}
\value{
The input object with all \code{NA} values replaced by \code{""}.
}
\description{
Converts all \code{NA} elements in a vector or matrix to \code{""}.
}
