% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.R
\name{OR.NA.to.F}
\alias{OR.NA.to.F}
\title{Convert missing values to FALSE}
\usage{
OR.NA.to.F(x)
}
\arguments{
\item{x}{Logical vector or matrix.}
}
\value{
The input object with all \code{NA} values replaced by \code{FALSE}.
}
\description{
Converts all \code{NA} elements in a vector or matrix to \code{FALSE}.
}
\seealso{
Other basic: 
\code{\link{OR.0.rm}()},
\code{\link{OR.F.to.NA}()},
\code{\link{OR.NA.rm}()},
\code{\link{OR.NA.to.0}()},
\code{\link{OR.NA.to.empty}()},
\code{\link{OR.as.numeric}()},
\code{\link{OR.string.to.binary}()}
}
\concept{basic}
