% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.R
\name{OR.NA.to.F}
\alias{OR.NA.to.F}
\title{Convert missing values to FALSE}
\usage{
OR.NA.to.F(x)
}
\arguments{
\item{x}{Numeric vector or matrix.}
}
\value{
The input object with all \code{NA} values replaced by \code{FALSE}.
}
\description{
Converts all \code{NA} elements in a vector or matrix to \code{FALSE}.
}
