% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.other.R
\name{OR.permutations}
\alias{OR.permutations}
\title{Enumerate permutations of k out of n items}
\usage{
OR.permutations(n, k)
}
\arguments{
\item{n}{Integer size of the set.}

\item{k}{Integer number of elements per permutation.}
}
\value{
Integer matrix with \code{factorial(n)/factorial(n-k)} rows and \code{k} columns.
}
\description{
Returns all length-\code{k} permutations of the integers from 1 to \code{n}, in
lexicographic order.
}
\seealso{
Other other: 
\code{\link{OR.collapse}()},
\code{\link{OR.holdout}()},
\code{\link{OR.read.csv.stub}()}
}
\concept{other}
