% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.R
\name{OR.rowmax}
\alias{OR.rowmax}
\title{Row-wise maxima}
\usage{
OR.rowmax(input)
}
\arguments{
\item{input}{A numeric matrix or data frame.}
}
\value{
A numeric vector of row-wise maxima.
}
\description{
Computes the maximum value of each row in a matrix or data frame, ignoring
\code{NA}s. Returns \code{NA} if an entire row is missing.
}
\details{
An example use case in survival analysis is determining the date of last
follow-up from several dates when the patient was observed.
}
