% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.R
\name{OR.string.to.binary}
\alias{OR.string.to.binary}
\title{Convert string variable to binary variable}
\usage{
OR.string.to.binary(input, pattern, missing)
}
\arguments{
\item{input}{String vector to evaluate.}

\item{pattern}{String vector of fixed patterns to each within \code{input}.}

\item{missing}{String vector of exact values to treat as \code{NA}.}
}
\value{
Numeric vector of \code{1} (pattern matched), \code{0} (no match), or \code{NA}
(missing).
}
\description{
Converts a string vector to a numeric vector according to the following:
\enumerate{
\item Elements exactly matching any value in \code{missing} return \code{NA}.
\item Otherwise, elements matching any fixed pattern in \code{pattern} return \code{1}.
\item All other elements return \code{0}.
}
}
\examples{
print(OR.string.to.binary(c("Yes", "Yes*", "No", "Missing"), "Yes", "Missing"))
}
