% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.summary.R
\name{OR.sum}
\alias{OR.sum}
\title{Sum with NA removal or NA if no valid entries}
\usage{
OR.sum(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Sum of non-\code{NA} elements or \code{NA} if none exist.
}
\description{
Equivalent to \code{sum(x, na.rm = TRUE)} except returns \code{NA} (rather than \code{0}) if
all elements are \code{NA}.
}
\seealso{
Other summary: 
\code{\link{OR.max}()},
\code{\link{OR.mean}()},
\code{\link{OR.min}()},
\code{\link{OR.rowleft}()},
\code{\link{OR.rowmax}()},
\code{\link{OR.rowmax.threshold}()},
\code{\link{OR.rowmin}()},
\code{\link{OR.rowmin.threshold}()}
}
\concept{summary}
