% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncReg.R
\name{OR.survoutcome}
\alias{OR.survoutcome}
\title{Survival outcomes}
\usage{
OR.survoutcome(date_start, date_event, date_follow, divisor = 365.2425/12)
}
\arguments{
\item{date_start}{Vector of start dates.}

\item{date_event}{Vector of event dates.}

\item{date_follow}{Vector of last follow-up dates for censored cases.}

\item{divisor}{Unit conversion factor for time. Default = \code{365.2425/12}.}
}
\value{
A numeric matrix with two columns: survival time and status (\code{1} =
event, \code{0} = censored). Cases with missing start dates or with survival time
≤ 0 are returned as \code{NA} in both columns.
}
\description{
Computes event times and statuses from start, event, and review dates.
}
