% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlm.R
\name{rlm.Huber}
\alias{rlm.Huber}
\title{Robust linear regression using Huber M-estimation}
\usage{
rlm.Huber(X, y, k2 = 1.345, maxit = 20, tol = 1e-04)
}
\arguments{
\item{X}{numeric design matrix.}

\item{y}{numeric response vector.}

\item{k2}{tuning constant for Huber psi. Default = \code{1.345}.}

\item{maxit}{maximum number of iterations. Default = \code{20}.}

\item{tol}{convergence tolerance. Default = \code{0.0001}.}
}
\value{
An object of class \code{"rlm"} and \code{"lm"}, broadly compatible with objects
returned by \code{MASS::rlm()} with minor differences due to rounding.
}
\description{
Fits a linear regression model using Huber M-estimation via iteratively
reweighted least squares.
}
